+++ src/execs/create_table_exec.hpp
@@ -41,7 +41,7 @@ public:
     throw NotImplementedException("next is not implemented for CreateTableExec!");
   }
 
-  const string message_on_completion(UNUSED size_t result_count) const override {
+  const string message_on_completion(UNUSED int32_t result_count) const override {
     return "Created table " + plan_->table_name();
   }
