+++ src/execs/delete_exec.hpp
@@ -38,7 +38,7 @@ public:
     return exec_ctx_.catalog().table_heap_for(table_oid);
   }
 
-  const string message_on_completion(size_t result_count) const override {
+  const string message_on_completion(int32_t result_count) const override {
     return "Deleted " + std::to_string(result_count) + " record(s)";
   }
