+++ src/execs/limit_exec.hpp
@@ -27,7 +27,7 @@ class LimitExec : public BaseExec {
     return Tuple();
   }
 
-  const string message_on_completion(size_t result_count) const override {
+  const string message_on_completion(int32_t result_count) const override {
     return "Found " + std::to_string(result_count) + " record(s)";
   }
