+++ src/execs/update_exec.hpp
@@ -13,7 +13,7 @@ public:
   bool has_next() override;
   Tuple next() override;
 
-  const string message_on_completion(size_t result_count) const override {
+  const string message_on_completion(int32_t result_count) const override {
     return "Updated " + std::to_string(result_count) + " record(s)";
   }
