+++ src/opers/seq_scan_oper.hpp
@@ -22,7 +22,7 @@ public:
     return true;
   }
 
-  const string message_on_completion(size_t result_count) const override {
+  const string message_on_completion(int32_t result_count) const override {
     return "Found " + std::to_string(result_count) + " record(s)";
   }
